requirements.txt
streamlit
groq
pymongo
python-dotenv
requests
beautifulsoup4
fuzzywuzzy
python-levenshtein
streamlit-searchbox
streamlit-keyup
trafilatura
newsapi-python